refactor: simplified logic
