+++ packages/renderer/src/lib/kube/resource-permission.ts
@@ -26,6 +26,7 @@ export async function listenResourcePermitted(
   callback: (permitted: boolean) => void,
 ): Promise<IDisposable> {
   const experimental = await window.isExperimentalConfigurationEnabled('kubernetes.statesExperimental');
+
   let contextName = '';
   let permissions: ContextPermission[] = [];
