+++ packages/renderer/src/lib/statusbar/StatusBar.spec.ts
@@ -126,6 +126,7 @@ describe('providers', () => {
   test('providers are hidden when configuration changes from true to false', async () => {
     vi.mocked(window.isExperimentalConfigurationEnabled).mockResolvedValueOnce(false);
     vi.mocked(window.isExperimentalConfigurationEnabled).mockResolvedValueOnce(true);
+
     render(StatusBar);
 
     await vi.waitFor(() => {
