+++ packages/renderer/src/stores/kubernetes-experimental.spec.ts
@@ -33,7 +33,7 @@ test('experimental mode is set', async () => {
   });
 });
 
-test('experimental mode is set to false', async () => {
+test('experimental mode is not set', async () => {
   vi.mocked(window.isExperimentalConfigurationEnabled).mockImplementation(async () => {
     return false;
   });
@@ -45,7 +45,7 @@ test('experimental mode is set to false', async () => {
 });
 
 test('experimental mode is not defined', async () => {
-  vi.mocked(window.isExperimentalConfigurationEnabled).mockRejectedValue(new Error('an error'));
+  vi.mocked(window.getConfigurationValue).mockRejectedValue(new Error('an error'));
   configurationProperties.set([]);
   await vi.waitFor(() => {
     const result = get(isKubernetesExperimentalModeStore);
